Added Roster Buttons
Added ability to quickly check current team rosters with a button on each list element. Needed to update CSS to match and added a bit to compensate for Bootstrap bug. Also added some CSS for anternate list formatting that I might switch to with owl-teams CSS class.

And cleaned up some code that was unnecessary or confusing.